requirements.txt
# Python 3.10–3.12 recommended

# FastAPI and web framework
fastapi==0.111.0
uvicorn[standard]==0.32.1
python-multipart==0.0.20
aiofiles==24.1.0

# Database and ORM
sqlalchemy==2.0.36
alembic==1.14.1
# Postgres driver (use psycopg2-binary on Linux/Mac, asyncpg on Windows)
psycopg2-binary==2.9.9; platform_system != "Windows"
asyncpg==0.29.0; platform_system == "Windows"

pgvector==0.4.1

# Redis and job queues
redis==5.2.1
rq==1.16.1

# Pydantic and validation
pydantic==2.10.4
pydantic-settings==2.6.1
pydantic-extra-types==2.8.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Document processing
pypdf==4.1.0
python-docx==1.1.2

# Agentic & LangChain stack (align on 0.3.x family for compatibility)
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.31
langgraph==0.6.6

# LLM provider for CV evaluation
groq==0.4.2

# Vector embeddings and search
sentence-transformers==3.0.1
# NOTE: faiss-cpu wheels can be flaky on Windows; skip there and use pgvector only.
faiss-cpu==1.8.0; platform_system != "Windows"

# HTTP client — pin <0.28 due to breaking changes used by OpenAI/LangChain clients
httpx==0.27.2

# Monitoring and observability (align instrumentation 0.33b0 with API/SDK 1.12.x)
sentry-sdk[fastapi]==2.17.0
opentelemetry-api==1.12.0
opentelemetry-sdk==1.12.0
opentelemetry-semantic-conventions==0.33b0
opentelemetry-instrumentation==0.33b0
opentelemetry-instrumentation-fastapi==0.33b0
opentelemetry-instrumentation-sqlalchemy==0.33b0

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
schemathesis==3.39.16 

# Development and code quality
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0

# Utilities
python-dotenv==1.0.1


# Scientific / data stack
numpy>=2.1.0

scipy==1.14.1
pandas==2.2.3

# Parsing / NLP
beautifulsoup4==4.12.3
lxml==5.3.0
nltk==3.9.1
spacy==3.7.5; platform_system != "Windows"

# Utilities
werkzeug==3.0.4
jinja2==3.1.4

# Testing
coverage==7.6.1

